next

+++ src/main/java/me/catzy/prestiz/objects/instruktorzy/InstruktorzyRepository.java
@@ -4,11 +4,11 @@
 
 package me.catzy.prestiz.objects.instruktorzy;
 
-import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "instruktorzy", path = "instruktorzy")
-public interface InstruktorzyRepository extends JpaRepository<Instruktor, Integer>
+public interface InstruktorzyRepository extends JpaRepository<Instruktor, Long>
 {
     Instruktor getById(final int id);
